Add clean working directory check to build scripts

+++ build-http.sh
@@ -2,6 +2,12 @@
 
 echo Building HTTP
 
+if [ -n "$(git status --untracked-files=no --porcelain)" ]
+then
+  echo -- Git working directory not clean
+  exit 1
+fi
+
 rm -rf build/http
 mkdir -p build/http
